+++ lib/screens/auth/login.dart
@@ -5,8 +5,8 @@ import 'package:provider/provider.dart';
 
 import '../../core/models/user.dart';
 import '../../core/providers/user.dart';
-import '../../core/routes/router.dart';
-import '../../core/routes/routes.dart';
+import '../../routes/router.dart';
+import '../../routes/routes.dart';
 import '../../styles/ui/colors.dart';
 import 'onboarding.dart';
