+++ lib/screens/auth/register.dart
@@ -8,8 +8,8 @@ import 'package:provider/provider.dart';
 
 import '../../core/models/user.dart';
 import '../../core/providers/user.dart';
-import '../../core/routes/router.dart';
-import '../../core/routes/routes.dart';
+import '../../routes/router.dart';
+import '../../routes/routes.dart';
 import '../../styles/icons/chap_chap_icons.dart';
 import '../../styles/ui/colors.dart';
 import 'login.dart';
