+++ lib/screens/auth/resetPassword.dart
@@ -2,12 +2,11 @@ import 'package:client/core/providers/user.dart';
 import 'package:client/screens/auth/login.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-
-import 'package:client/core/routes/routes.dart';
-import 'package:client/core/routes/router.dart';
-import 'package:client/styles/ui/colors.dart';
 import 'package:provider/provider.dart';
 
+import '../../routes/routes.dart';
+import '../../routes/router.dart';
+import '../../styles/ui/colors.dart';
 import 'onboarding.dart';
 
 class ResetPassword extends StatelessWidget {
