+++ lib/screens/roles/admin/home.dart
@@ -1,6 +1,3 @@
-import 'package:client/core/providers/appdata.dart';
-import 'package:client/core/router/roles.dart';
-import 'package:client/styles/ui/colors.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:dotted_border/dotted_border.dart';
 import 'package:flutter/material.dart';
@@ -8,6 +5,9 @@ import 'package:provider/provider.dart';
 
 import '../../../core/models/user.dart';
 import '../../../styles/icons/chap_chap_icons.dart';
+import '../../../core/providers/appdata.dart';
+import '../../../router/roles.dart';
+import '../../../styles/ui/colors.dart';
 import 'items.dart';
 
 class AdminHome extends StatelessWidget {
