+++ lib/screens/shared/profile.dart
@@ -1,4 +1,3 @@
-import 'package:client/core/router/roles.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -6,6 +5,7 @@ import 'package:provider/provider.dart';
 
 import '../../core/providers/user.dart';
 import '../../styles/ui/colors.dart';
+import '../../router/roles.dart';
 
 class ProfilePage extends StatelessWidget {
   ProfilePage({Key? key}) : super(key: key);
